requirements.txt
faust-streaming==0.7.8
jsonlines==3.0.0
kafka-python==2.0.2